add activity info and edit fragment pages adapter
