--- Onews-mobile-android/Onews/app/src/main/java/com/example/dongson/onews/view/ContentFragment.java
@@ -1,30 +1,0 @@
-package com.example.dongson.onews.view;
-
-import android.content.Context;
-import android.graphics.Color;
-import android.net.Uri;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.FrameLayout;
-import android.widget.TextView;
-
-import com.example.dongson.onews.R;
-
-
-public class ContentFragment extends Fragment {
-
-    public ContentFragment() {
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_content, container, false);
-        TextView textView = (TextView) rootView.findViewById(R.id.section_label);
-        textView.setText("dmm");
-        return rootView;
-    }
-}
