+++ Onews-mobile-android/Onews/app/src/main/java/com/example/dongson/onews/view/MainFragment.java
@@ -28,8 +28,8 @@ public class MainFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View rootView = inflater.inflate(R.layout.fragment_main, container, false);
-        TextView textView = (TextView) rootView.findViewById(R.id.section_label);
-        textView.setText("Kteam");
+//        TextView textView = (TextView) rootView.findViewById(R.id.section_label);
+//        textView.setText("Kteam");
         return rootView;
     }
 }
